Implementing shopping cart

+++ app/src/main/java/com/allysonjeronimo/marvelapp/data/db/entity/Comic.kt
@@ -14,5 +14,5 @@ data class Comic(
     val thumbnailExtension:String,
     val price:Double,
     val creators:String,
-    val issueNumber:Double
+    val issueNumber:Int
 )
